move test model into its own models.py so it can be found by the test runner
